app/src/main/java/com/example/sudoku/Word.java
package com.example.sudoku;

import androidx.room.Entity;
import androidx.room.PrimaryKey;

@Entity(tableName = "word_table")
public class Word {

    @PrimaryKey(autoGenerate = true)
    private int id;

    private String english;

    private String french;

    private String name;

    private int incorrectCount;

    public Word(String english, String french, String name) {
        this.english = english;
        this.french = french;
        this.name = name;
        this.incorrectCount = 0;
    }

    public void setId(int id) {
        this.id = id;
    }

    public int getId() {
        return id;
    }

    public String getEnglish() {
        return english;
    }

    public String getFrench() {
        return french;
    }

    public String getName() {
        return name;
    }

    public int getIncorrectCount() {
        return incorrectCount;
    }

    public void setIncorrectCount(int incorrectCount) {
        this.incorrectCount = incorrectCount;
    }
}
